Move coordinatortest to pkg/coordinator/internal
