+++ pkg/coordinator/internal/test/testdata/find_deposits_scenario_0.json
