+++ pkg/coordinator/internal/test/testdata/find_deposits_scenario_1.json
