+++ pkg/coordinator/internal/test/testdata/find_deposits_scenario_3.json
