+++ pkg/coordinator/internal/test/testdata/propose_sweep_scenario_1.json
